Complete basic functionality
Buttons have been given event listeners. Number
buttons will print to the display. An operator
button will take what's in the display as value 1
and itself as an operator and clears the display
to add a new value. Then the equals button takes
whatever new value is in the display as the second
value and passes those to the operate function.