Fix the answer concatenation bug
When the user did a sum and pressed equals the
display would show the answer. Any subsequent
number pressed would concatenate to the answer
instead of starting a new calculation. I've done
this by forcing a 'clear' click when the display
shows the answer and the answer variable isn't
empty, so it doesn't clear on the first sum done.