Made small display print current running calc
I've changed the chained calculations to also show
the operator used. Which meant I had to change the
equals and operator operate clauses to omit the
operator first, as they would take the current
total value from the small display which would
result in NaN because of the added operator in the
string.