Change the display to an input box
In an attempt to limit the amount of characters
the calculator will accept, I changed the display
from a div to an input box. I've disabled it and
edited the js so as to properly print to the box
as it did with the div. The issue is, the field
does stop printing numbers, but the calculator
still takes in these bigger numbers, you just
can't see it now.

I think I'll have to create some kind of listener
to limit the size of the number.