Build % function
I've built the % button function that will find
percentages of numbers, will add or subtract them
from the original number, or give errors if used
improperly. I had to add another clause to the
operator function if statement to allow for the
answer of a percent check, to be chained into
other sums.