Added minus first
Edited the operator function to allow for the
minus button to be pressed first to start with a
negative number.

I also trialed a bracket button, started building
the function but I've commented it out and I'll be
leaving it in in case I come back to it. I can't
really find a use case for it at the moment.