Conjoin click listeners
Instead of having individual click event listeners
for all the buttons aside from the numbers and
operators, I've conjoined them into a couple new
shared event listeners, using if statements to
parse which was clicked.

It's allowed me to delete multiple declarations at
the top of the script and cleaned up the code.