Potential bug fix
Changed the first else if in the second else if of
the operator event listener. I had an error after
a sum chaining into a / sum using the previous
answer. It didn't need to operate, it was mostly
reordering the lines to make sure it did things in
the right order.