Change container measurement units
I've changed the sizing of the container that
houses the entire calculator to vh and vw instead
of px. I was interested on how to make it look
better on an android browser, so I'm testing out a
newer size. I also had to slightly change the
width of the equals button as it was set via px,
it is now flex 2.2 which seems to be about the
right size.