Bug fixing
There is a bug during initial sums when argument
two is empty that's difficult to fix. At the
moment I've fixed it to the point that you end up
having to press a button twice if you're trying to
add a double number like 22 or 44 or 55 to
something. The initial number press will change
the number (unless it's the same number as the
display) but the second press, if it's the same as
the display number, won't register, it will change
the argumentTWo so that the next time you press it
it will add it to the display. So it's workable in
this current fix, just not great. Other fixes I've
tried make it break in other circumstances.