repeating number workaround
There is a problem with the display showing the
current value and not emptying the display before
entering your second argument. So I commented out
that code and added the clear display to the
operator buttons.

There is still a bug in which if I do 1+2+3+ it
won't add the 3. I'm assuming because the small
display shows 3 but 2+2+2+ works. So it's some
edge case I still have to find.